make activity timestamps timezone-aware
Hamster deals only in "naive" datetime objects (that is, with
tzinfo=None). They are always in local time. With this patch the
TimeEntry.datetime attribute (which is populated from Hamster's
Fact.start_time) has its timezone set to dateutil.tz.tzlocal() which
makes it easier when dealing with external services that are
timezone-aware and need timezone offsets in their timestamps.