allow using a custom query to map tickets to JIRA issues
This is useful if you're using JIRA to also track tickets from another
system (for example Bugzilla). Ticket values resembling a JIRA issue key
(PRJ-123) will be used as is, but other ticket values will be run
through the configured JQL query. The query is expected to return
a single JIRA issue.

Here is an example, for a JIRA project that is using custom field 10400
to track external bug URLs:

    [profile "myjira"]
    format = jira
    url = https://jira.example.com/
    ticket_jql_format = cf[10400] = 'https://bugzilla.example.com/show_bug.cgi?id={}'

With this configuration you could track activity in Hamster against both
JIRA issue keys and Bugzilla bug IDs.